Add files via upload
Challenge Description --> Write a function that returns the number of prime numbers that exist up to and including a given number. (e.g. if 100.25 is passed in, it should return 25, because there are 25 prime numbers up to and including 100.25)

What makes my approach good? This algorithm is designed with micro-optimizations that substantially lower the number of calculations needed to return the value. This not only offers a tremendous increase in speed, but reduces the memory requirement for the algorithm, allowing large numbers to be calculated quickly. I tested it using 

There are other algorithms online that perform this task quickly, but though I found most of them difficult to follow along with, so I made it a goal to make mine as easy to follow along with as possible.

The function only bothers to check odd numbers, since no even number other than 2 is prime. I have accounted for 2 by assuming there to be at least one prime. If the function takes in 2 it returns a prime count of 1, and if it is less than two, it returns 0.

For the numbers it does check, it will only check multipliers up to the square root of the number (with decimals truncated), since the equations will mirror each other after the root of the number.

E.G. the factors of 16 can look like this:
1 x 16
2 x 8
4 x 4
8 x 2
16 x 1

Obviously, my function would never bother with 16, since it's an even number, but the point is that the possible factors of any number will be the same after you hit the square root of the number, so I made the square root of the value being checked to be the upper limit of the range of numbers generated to calculate. I also made sure to add a break statement, so I don't waste any more processing time on a number that has already been shown to be composite.