Add files via upload
Often this algorithm is done recursively, which uses a colossal amount of processing power relative to the function written here. The function I wrote here only ever stores the last two numbers in memory, which is the minimum amount possible to calculate Fibonacci numbers. I created a two-element-long list which drops the lower number in each calculation and replaces it with the next value in the sequence. As per usual, I keep my variables descriptive and my code clean.